Relocate utility classes to core.utils package.
Moved `FormatRates.kt`, `LanguageManager.kt`, and `ThemeManager.kt` from `com.currency.rateman.utils` to `com.currency.rateman.core.utils`.
Updated import statements in affected files (`SettingsViewModel.kt`, `ProviderDetailScreen.kt`, `MainActivity.kt`, `ProviderItem.kt`) to reflect the new package structure.
