+++ app/src/main/java/com/currency/rateman/MainActivity.kt
@@ -8,8 +8,8 @@ import androidx.activity.enableEdgeToEdge
 import androidx.lifecycle.lifecycleScope
 import com.currency.rateman.ui.navigation.AppRouter
 import com.currency.rateman.ui.theme.RateManAppTheme
-import com.currency.rateman.utils.LanguageManager
-import com.currency.rateman.utils.ThemeManager
+import com.currency.rateman.core.utils.LanguageManager
+import com.currency.rateman.core.utils.ThemeManager
 import kotlinx.coroutines.launch
 
 class MainActivity : ComponentActivity() {
