+++ app/src/main/java/com/currency/rateman/core/utils/FormatRates.kt
@@ -1,4 +1,4 @@
-package com.currency.rateman.utils
+package com.currency.rateman.core.utils
 
 fun Double.formatRate(): String {
     return this.toString().trimEnd('0').trimEnd('.')
