+++ app/src/main/java/com/currency/rateman/provider/ui/components/ProviderItem.kt
@@ -18,7 +18,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import com.currency.rateman.core.data.model.CurrencyRate
 import com.currency.rateman.provider.data.model.Provider
-import com.currency.rateman.utils.formatRate
+import com.currency.rateman.core.utils.formatRate
 
 @Composable
 fun ProviderItem(
