+++ app/src/main/java/com/currency/rateman/ui/viewmodels/SettingsViewModel.kt
@@ -8,8 +8,8 @@ import com.currency.rateman.core.data.model.enums.CurrencyCode
 import com.currency.rateman.core.data.model.enums.LanguageCode
 import com.currency.rateman.core.data.model.enums.ThemeMode
 import com.currency.rateman.core.data.repository.SettingsRepository
-import com.currency.rateman.utils.LanguageManager
-import com.currency.rateman.utils.ThemeManager
+import com.currency.rateman.core.utils.LanguageManager
+import com.currency.rateman.core.utils.ThemeManager
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
